simplemind/requirements/deep_med_requirements.txt
h5py==2.10.0
Keras==2.3.1
matplotlib==3.2.1
MedPy==0.4.0
#numpy==1.18.2
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.1
#PyYAML==5.3.1
setuptools
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
sklearn==0.0
# tensorflow-gpu==1.14.0
tensorboard==1.15.0
virtualenv==20.0.17

# 05012021 after Youngwon's integration of CNNs to MIU
pydicom==2.1.2
tensorflow-gpu==1.15.0

# 11042021 after Youngwon's integration of file locking system to MIU
psutil

# 08172022 for report generation
jinja2
markupsafe==2.0.1   #needed for jinja2 to work

# 08192022 for GA 
deap==1.3.1
